# Speed up GTG Simulator by reducing slices/matches