--- a/actionpack/test/journey/gtg/builder_test.rb
+++ b/actionpack/test/journey/gtg/builder_test.rb
@@ -8,13 +8,13 @@ module ActionDispatch
       class TestBuilder < ActiveSupport::TestCase
         def test_following_states_multi
           table = tt ["a|a"]
-          assert_equal 1, table.move([[0, nil]], "a", 0, 1).length
+          assert_equal 1, table.move([[0, nil]], "a", "a", 0, true).length
         end
 
         def test_following_states_multi_regexp
           table = tt [":a|b"]
-          assert_equal 1, table.move([[0, nil]], "fooo", 0, 4).length
-          assert_equal 2, table.move([[0, nil]], "b", 0, 1).length
+          assert_equal 1, table.move([[0, nil]], "fooo", "fooo", 0, true).length
+          assert_equal 2, table.move([[0, nil]], "b", "b", 0, true).length
         end
 
         def test_multi_path
@@ -26,7 +26,7 @@ module ActionDispatch
             [2, "/"],
             [1, "c"],
           ].inject([[0, nil]]) { |state, (exp, sym)|
-            new = table.move(state, sym, 0, sym.length)
+            new = table.move(state, sym, sym, 0, sym != "/")
             assert_equal exp, new.length
             new
           }