@byroot 2025-02-11
I rewrote your benchmark to better see the difference:

```
== index ==
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
               after    29.146k i/100ms
Calculating -------------------------------------
               after    292.018k (± 0.8%) i/s    (3.42 μs/i) -      1.486M in   5.090560s

Comparison:
              before:   273096.8 i/s
               after:   292018.2 i/s - 1.07x  faster

== show ==
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
               after    22.407k i/100ms
Calculating -------------------------------------
               after    224.862k (± 0.3%) i/s    (4.45 μs/i) -      1.143M in   5.082102s

Comparison:
              before:   196157.8 i/s
               after:   224861.5 i/s - 1.15x  faster

== show_nested ==
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
               after    19.688k i/100ms
Calculating -------------------------------------
               after    196.771k (± 0.2%) i/s    (5.08 μs/i) -    984.400k in   5.002789s

Comparison:
              before:   171470.6 i/s
               after:   196770.8 i/s - 1.15x  faster
```

```ruby
# frozen_string_literal: true

require "bundler/inline"
gemfile do
  gem "rails", path: "/Users/byroot/src/github.com/byroot/rails"
  gem "benchmark-ips"
end

require "action_dispatch"
require "benchmark/ips"

routes = ActionDispatch::Routing::RouteSet.new
routes.draw do
  get "/products", to: ->(e) { [200, {}, ["p"]] }
  get "/products/:id", to: ->(e) { [200, {}, ["pid"]] }
  get "/collections", to: ->(e) { [200, {}, ["c"]] }
  get "/collections/:id/products", to: ->(e) { [200, {}, ["cidp"]] }
end

stage = ENV.fetch("STAGE", "before")

requests = {
  index: {
    "REQUEST_METHOD" => "GET",
    "PATH_INFO" => "/products",
    "SCRIPT_NAME" => "",
    "rack.input" => File.open("/dev/null"),
  }.freeze,

  show: {
    "REQUEST_METHOD" => "GET",
    "PATH_INFO" => "/products/1",
    "SCRIPT_NAME" => "",
    "rack.input" => File.open("/dev/null"),
  }.freeze,

  show_nested: {
    "REQUEST_METHOD" => "GET",
    "PATH_INFO" => "/collections/1/products",
    "SCRIPT_NAME" => "",
    "rack.input" => File.open("/dev/null"),
  }.freeze,
}.each do |name, env|
  puts "== #{name} =="
  Benchmark.ips do |x|
    x.report(stage) { routes.call(env.dup) }
    x.save!("/tmp/action_dispatch_#{name}")
    x.compare!(order: :baseline)
  end
end
```