@byroot 2025-02-11
Not sure if worth it, but I can optimize this farther by using that escape table all the way: https://github.com/byroot/rails/commit/8545fbd8c3c54bf6e1eef01fc8338198678c52dc

```
```
== index ==
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
               after    33.759k i/100ms
Calculating -------------------------------------
               after    338.664k (± 0.4%) i/s    (2.95 μs/i) -      1.722M in   5.083918s

Comparison:
              before:   269469.3 i/s
               after:   338664.4 i/s - 1.26x  faster

== show ==
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
               after    25.635k i/100ms
Calculating -------------------------------------
               after    257.573k (± 0.3%) i/s    (3.88 μs/i) -      1.307M in   5.075826s

Comparison:
              before:   194468.2 i/s
               after:   257572.6 i/s - 1.32x  faster

== show_nested ==
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
               after    22.175k i/100ms
Calculating -------------------------------------
               after    222.257k (± 2.4%) i/s    (4.50 μs/i) -      1.131M in   5.091618s

Comparison:
              before:   165457.3 i/s
               after:   222256.8 i/s - 1.34x  faster
```

A ~30% gain is nice, but it's 30% of something that's already not much (6μs on my machine). So I guess it's mostly a question of code readability.