@skipkayhil 2025-02-11
> I rewrote your benchmark to better see the difference:

:bow: I was trying to figure out how to make that work, thanks for the tip

> Not sure if worth it, but I can optimize this farther by using that escape table all the way: [byroot@8545fbd](https://github.com/byroot/rails/commit/8545fbd8c3c54bf6e1eef01fc8338198678c52dc)

Ah, that makes a lot of sense! Applied.

> A ~30% gain is nice, but it's 30% of something that's already not much (6μs on my machine). So I guess it's mostly a question of code readability.

I recognize its a microbenchmark, but I see the TechEmpower benchmarks mentioned a lot and especially since [many of the routes](https://github.com/TechEmpower/FrameworkBenchmarks/blob/e38f75a538ad056426e06968cc9def9e5a06c0e7/frameworks/Ruby/rails/config/routes.rb#L9) skip Action Controller and/or Action View I think the ~25% speedup will translate pretty directly there.