@byroot 2025-02-12
> I recognize its a microbenchmark, but I see the TechEmpower benchmark mentioned a lot

Ah, so the goal is to go ROFLscale. If that's so there some more nasty tricks that can be pulled, e.g. flattening the transition tables: https://github.com/byroot/rails/commits/opt-gtg-move/, which speed your benchmark by another 10%.

There's probably some other tricks to pull.