@byroot 2025-02-12
There is also https://github.com/rails/rails/commit/c1c528d9f26195d08028caa0a8cc8bee4e14d962, which gives another 5%, but I think the Visitor could be optimized much more by using code generation.